loader: use hbloader to load NROs
